bughound_api/requirements.txt
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
mysqlclient==1.4.1
Pillow==5.4.1
django-filter==2.1.0
djangorestframework-filters==0.10.2.post0
